requirements.txt
Django==2.2
django-sslserver
djangorestframework
psycopg2-binary
dj-database-url